+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KN91R29JF6ZFW5YM1GWFRAG0"
+    id: Optional[str] = "01KN9KG40PB2ETJ416CW9T008D"
 
     description: Optional[str] = None